requirements.txt
streamlit==1.29.0
pandas==2.1.0
nltk==3.8.1
gpt4all==2.0.2
langchain==0.0.27
langchain-community==0.0.13
langchain-huggingface==0.0.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.24.3